requirements.txt
qrcode
pillow
base58
segno
qrcodegen
pycryptodome
